backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.27.0
requests==2.31.0
aiohttp==3.9.5
tree-sitter==0.20.4
tree-sitter-languages==1.9.1
networkx==3.2.1
gitpython==3.1.40
python-multipart==0.0.6
python-dotenv==1.0.0
asyncpg==0.29.0
pymongo==4.6.1
pymilvus==2.3.4
pyjwt[crypto]==2.8.0
upstash-redis>=1.0.0
openai>=1.0.0